pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "security-use"
version = "0.2.9"
description = "Security scanning tool for dependencies and Infrastructure as Code"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Security Use", email = "security@example.com" }
]
keywords = ["security", "vulnerability", "scanner", "dependencies", "iac", "terraform", "cloudformation"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security",
    "Topic :: Software Development :: Quality Assurance",
]

dependencies = [
    "httpx>=0.25.0",
    "packaging>=23.0",
    "python-hcl2>=4.3.0",
    "pyyaml>=6.0",
    "rich>=13.0.0",
    "click>=8.0.0",
    "tomli>=2.0.0;python_version<'3.11'",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
    "types-PyYAML>=6.0.0",
]
sensor = [
    "fastapi>=0.100.0",
    "flask>=2.0.0",
]

[project.scripts]
security-use = "security_use.cli:main"

[project.urls]
Homepage = "https://github.com/security-use/security-use"
Repository = "https://github.com/security-use/security-use"
Issues = "https://github.com/security-use/security-use/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/security_use"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_ignores = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
